requirements.txt
PyQt5
openai-whisper
torch
ffmpeg-python
numpy